+++ CHANGELOG.md
@@ -1,6 +1,30 @@
 LifterLMS Tests Changelog
 =========================
 
+v3.1.0 - 2021-10-04
+-------------------
+
++ Raise minimum supported PHP version to 7.3
++ Add support for phpunit 9.5
++ When running `llms-tests install`, the default value of `WP_TESTS_VERSION` is now set to `trunk`.
+
+
+v3.0.0 - 2021-09-27
+-------------------
+
++ Updated to rely on `yoast/phpunit-polyfills`.
++ Renamed method `LLMS_Unit_Test_Case_Base::setUp()` to `LLMS_Unit_Test_Case_Base::set_up()` for WP core compat.
++ Renamed method `LLMS_Unit_Test_Case_Base::tearDown()` to `LLMS_Unit_Test_Case_Base::tear_down()` for WP core compat.
++ Renamed method `LLMS_REST_Unit_Test_Case::setUp()` to `LLMS_REST_Unit_Test_Case::set_up()` for WP core compat.
++ Renamed method `LLMS_REST_Unit_Test_Case::tearDown()` to `LLMS_REST_Unit_Test_Case::tear_down()` for WP core compat.
+
+
+v2.1.0 - 2021-08-26
+-------------------
+
++ Introduced pluggable version of `llms_exit()` which throws the `LLMS_Unit_Test_Exception_Exit` exception in favor of exiting.
+
+
 v2.0.2 - 2021-08-23
 -------------------